third_party/blink/web_tests/fast/dom/shadow/pseudo-host-parameter-matches-shadow-host-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test that :host(selector) can match a shadow host when the host matches the selector: http://crbug.com/313935.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
